Started Fix
Still having trouble with the code. In the program there is a checker on line 390 of Backroom that is giving me trouble. See, I will try to find the difference of the two messages, pad the privous one out and print the new one on the console. Sometimes, letters are left behind and I cant seem to figure out why.